Fix logtest integration tests
