+++ tests/integration/test_logtest/test_remove_session/test_remove_session.py
@@ -8,11 +8,11 @@ import pytest
 import yaml
 import json
 
-from wazuh_testing import global_parameters
-from wazuh_testing.analysis import callback_fim_error
 from wazuh_testing.tools import LOG_FILE_PATH, WAZUH_PATH
+from wazuh_testing.tools.monitoring import close_sockets
 
 # Marks
+from wazuh_testing.tools.monitoring import SocketController
 
 pytestmark = [pytest.mark.linux, pytest.mark.tier(level=0), pytest.mark.server]
 
@@ -32,10 +32,20 @@ receiver_sockets_params = [(logtest_path, 'AF_UNIX', 'TCP')]
 
 receiver_sockets, monitored_sockets, log_monitors = None, None, None  # Set in the fixtures
 
+
 def create_session():
-    receiver_sockets[0].send('{"event": "Jun 24 11:54:19 Master systemd[2099]: Started VTE child process 20118 launched by terminator process 17756.","log_format": "syslog","location": "master->/var/log/syslog"}', size=True)
-    result = json.loads(receiver_sockets[0].receive(size=True).rstrip(b'\x00').decode())["token"]
-    return result
+    receiver_sockets[0].send(
+        '{"event": "Jun 24 11:54:19 Master systemd[2099]: Started VTE child process 20118 launched by terminator process'
+        ' 17756.","log_format": "syslog","location": "master->/var/log/syslog"}',
+        size=True)
+    token = json.loads(receiver_sockets[0].receive(size=True).rstrip(b'\x00').decode())['token']
+
+    # Close socket
+    close_sockets(receiver_sockets)
+
+    # Renew socket for future connections
+    receiver_sockets[0] = SocketController(address=logtest_path, family='AF_UNIX', connection_protocol='TCP')
+    return token
 
 
 # Tests
@@ -58,11 +68,11 @@ def test_remove_session(connect_to_sockets_function, test_case: list):
         result = receiver_sockets[0].receive(size=True).rstrip(b'\x00').decode()
 
         assert stage['output'] == result, 'Failed test case stage {}: {}'.format(test_case.index(stage) + 1,
-                                                                                stage['stage'])
+                                                                                 stage['stage'])
     else:
         session_token = create_session()
         receiver_sockets[0].send(stage['input'].format(session_token), size=True)
         result = receiver_sockets[0].receive(size=True).rstrip(b'\x00').decode()
-
-        assert stage['output'].format(session_token) == result, 'Failed test case stage {}: {}'.format(test_case.index(stage) + 1,
-                                                                                stage['stage'])
+        assert stage['output'].format(session_token) == result, 'Failed test case stage {}: {}'.format(
+            test_case.index(stage) + 1,
+            stage['stage'])
